Split `hsd_397E` and `hsd_3982` out of `debugconsole_main`
`debugconsole_main.c`'s `.data` ended with two objects the compiler had
aligned to 8: the `baselib_mfspr` format string at 0x8040BF10 and the
`HSD_WObjDesc` at 0x8040BF48. MWCC never does that for a string or a
20-byte struct mid-object, so each one starts a translation unit.

Splitting there gives three objects whose `.text` and `.data` all match:
`debugconsole_main.c` (0x80394314), `hsd_397E.c` (`baselib_mfspr`), and
`hsd_3982.c` (`fn_803982E4`, `hsd_80398310`, and the camera descriptor).
`debugconsole_main` `.data` goes from 30.9% to 100%; the two new objects
are exact and linked.
