+++ config/GALE01/splits.txt
@@ -5723,11 +5723,19 @@ sysdolphin/baselib/hsd_393C.c:
 	.bss        start:0x804CF7E8 end:0x804CF810
 
 sysdolphin/baselib/debugconsole_main.c:
-	.text       start:0x80394314 end:0x803983A4
-	.data       start:0x8040AB00 end:0x8040BFB0
+	.text       start:0x80394314 end:0x80397E0C
+	.data       start:0x8040AB00 end:0x8040BF10
 	.bss        start:0x804CF810 end:0x804D08E8
 	.sdata      start:0x804D60B0 end:0x804D6358
 	.sbss       start:0x804D78C8 end:0x804D78D0
+
+sysdolphin/baselib/hsd_397E.c:
+	.text       start:0x80397E0C end:0x803982E4
+	.data       start:0x8040BF10 end:0x8040BF48
+
+sysdolphin/baselib/hsd_3982.c:
+	.text       start:0x803982E4 end:0x803983A4
+	.data       start:0x8040BF48 end:0x8040BFB0
 
 sysdolphin/baselib/particle.c:
 	extab       start:0x80005520 end:0x80005588
