+++ config/GALE01/symbols.txt
@@ -25947,8 +25947,9 @@ lbl_8040BEAC = .data:0x8040BEAC; // type:object size:0x17 scope:local data:strin
 lbl_8040BEC4 = .data:0x8040BEC4; // type:object size:0x1C scope:local
 lbl_8040BEE0 = .data:0x8040BEE0; // type:object size:0x14 data:string
 lbl_8040BEF4 = .data:0x8040BEF4; // type:object size:0x17 data:string
-lbl_8040BF10 = .data:0x8040BF10; // type:object size:0x32 scope:local data:string
+@73 = .data:0x8040BF10; // type:object size:0x32 scope:local data:string
 lbl_8040BF48 = .data:0x8040BF48; // type:object size:0x14 scope:local
+...data.0 = .data:0x8040BF48; // type:label scope:local
 lbl_8040BF5C = .data:0x8040BF5C; // type:object size:0x14 scope:local
 lbl_8040BF70 = .data:0x8040BF70; // type:object size:0x40 scope:local
 lbl_8040BFB0 = .data:0x8040BFB0; // type:object size:0xB scope:local data:string
