+++ src/melee/gm/gm_1601.c
@@ -43,6 +43,7 @@
 #include <sysdolphin/baselib/controller.h>
 #include <sysdolphin/baselib/debugconsole_main.h>
 #include <sysdolphin/baselib/hsd_3915.h>
+#include <sysdolphin/baselib/hsd_3982.h>
 #include <sysdolphin/baselib/random.h>
 #include <melee/gm/gm_1A45.h>
 #include <melee/pl/player.h>
