+++ src/sysdolphin/baselib/debugconsole_main.h
@@ -49,8 +49,5 @@
 /* 397DA4 */ void hsd_80397DA4(OSContext*);
 /* 397DF4 */ void Exception_StoreDebugLevel(int);
 /* 397DFC */ void hsd_80397DFC(u32 size);
-/* 397E0C */ int baselib_mfspr(int);
-/* 3982E4 */ void fn_803982E4(HSD_GObj*, int);
-/* 398310 */ HSD_GObj* hsd_80398310(u16, u8, u8, u32);
 
 #endif
