+++ src/sysdolphin/baselib/hsd_397E.h
@@ -1,0 +1,8 @@
+#ifndef SYSDOLPHIN_BASELIB_HSD_397E_H
+#define SYSDOLPHIN_BASELIB_HSD_397E_H
+
+#include <platform.h>
+
+/* 397E0C */ int baselib_mfspr(int);
+
+#endif
