+++ src/sysdolphin/baselib/hsd_3982.h
@@ -1,0 +1,11 @@
+#ifndef SYSDOLPHIN_BASELIB_HSD_3982_H
+#define SYSDOLPHIN_BASELIB_HSD_3982_H
+
+#include <platform.h>
+
+#include <sysdolphin/baselib/forward.h>
+
+/* 3982E4 */ void fn_803982E4(HSD_GObj*, int);
+/* 398310 */ HSD_GObj* hsd_80398310(u16, u8, u8, u32);
+
+#endif
